clean code updates / reorg
