--- .idea/.idea.DynamoApi/.idea/.name
@@ -1,1 +1,0 @@
-DynamoApi
