+++ .idea/.idea.Solar/.idea/.name
@@ -1,0 +1,1 @@
+Solar
