+++ Solar.sln
@@ -1,0 +1,106 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{F5B85E6F-DA69-4A6D-9F73-7A2DEF1A2BCB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solar.Web", "src\Solar.Web\Solar.Web.csproj", "{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{55ADF3A4-9090-43A2-A7FB-C45C61F47F07}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solar.IntegrationTests", "tests\Solar.IntegrationTests\Solar.IntegrationTests.csproj", "{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solar.UnitTests", "tests\Solar.UnitTests\Solar.UnitTests.csproj", "{64B7EEC4-E131-4E32-B041-5A97281A9559}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "solution-items", "solution-items", "{CB3B1DCC-1CB1-4113-AE45-8454149C1AB9}"
+ProjectSection(SolutionItems) = preProject
+	README.md = README.md
+EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solar.Core", "src\Solar.Core\Solar.Core.csproj", "{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solar.Infrastructure", "src\Solar.Infrastructure\Solar.Infrastructure.csproj", "{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Debug|x64.Build.0 = Debug|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Debug|x86.Build.0 = Debug|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Release|x64.ActiveCfg = Release|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Release|x64.Build.0 = Release|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Release|x86.ActiveCfg = Release|Any CPU
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376}.Release|x86.Build.0 = Release|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Debug|x64.Build.0 = Debug|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Debug|x86.Build.0 = Debug|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Release|x64.ActiveCfg = Release|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Release|x64.Build.0 = Release|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Release|x86.ActiveCfg = Release|Any CPU
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1}.Release|x86.Build.0 = Release|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Debug|x64.Build.0 = Debug|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Debug|x86.Build.0 = Debug|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Release|x64.ActiveCfg = Release|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Release|x64.Build.0 = Release|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Release|x86.ActiveCfg = Release|Any CPU
+		{64B7EEC4-E131-4E32-B041-5A97281A9559}.Release|x86.Build.0 = Release|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Debug|x64.Build.0 = Debug|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Debug|x86.Build.0 = Debug|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Release|x64.ActiveCfg = Release|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Release|x64.Build.0 = Release|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Release|x86.ActiveCfg = Release|Any CPU
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F}.Release|x86.Build.0 = Release|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Debug|x64.Build.0 = Debug|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Debug|x86.Build.0 = Debug|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Release|Any CPU.Build.0 = Release|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Release|x64.ActiveCfg = Release|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Release|x64.Build.0 = Release|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Release|x86.ActiveCfg = Release|Any CPU
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{EBC8F3F1-0DF9-4FB3-A45E-536B15DD7376} = {F5B85E6F-DA69-4A6D-9F73-7A2DEF1A2BCB}
+		{A2E477F2-4E0E-46BC-B2AF-653A38CEC5C1} = {55ADF3A4-9090-43A2-A7FB-C45C61F47F07}
+		{64B7EEC4-E131-4E32-B041-5A97281A9559} = {55ADF3A4-9090-43A2-A7FB-C45C61F47F07}
+		{F426B3B2-B278-4AF8-A6F6-FFBAEAA2FD1F} = {F5B85E6F-DA69-4A6D-9F73-7A2DEF1A2BCB}
+		{54A3DA8D-EAAE-4391-ADFE-FFE8444C1299} = {F5B85E6F-DA69-4A6D-9F73-7A2DEF1A2BCB}
+	EndGlobalSection
+EndGlobal
