--- src/DynamoApi/Controllers/VersionController.cs
@@ -1,25 +1,0 @@
-using System.Reflection;
-using Microsoft.AspNetCore.Mvc;
-
-namespace DynamoApi.Controllers
-{
-    [ApiController]
-    [Route("/version")]
-    public class VersionController
-    {
-        [HttpGet]
-        public IActionResult GetVersion()
-        {
-            // ReSharper disable once AssignNullToNotNullAttribute
-            var version = Assembly.GetEntryAssembly()
-                .GetCustomAttribute<AssemblyInformationalVersionAttribute>()
-                .InformationalVersion;
-            
-            var content = new
-            {
-                Version = version
-            };
-            return new OkObjectResult(content);
-        }
-    }
-}
