--- src/DynamoApi/HealthChecks/AliveHealthCheck.cs
@@ -1,16 +1,0 @@
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Diagnostics.HealthChecks;
-
-namespace DynamoApi.HealthChecks
-{
-    public class AliveHealthCheck : IHealthCheck
-    {
-        public Task<HealthCheckResult> CheckHealthAsync(
-            HealthCheckContext context,
-            CancellationToken cancellationToken = new CancellationToken())
-        {
-            return Task.FromResult(HealthCheckResult.Healthy("A healthy result."));
-        }
-    }
-}
