--- src/DynamoApi/Models/Moon.cs
@@ -1,9 +1,0 @@
-namespace DynamoApi.Models
-{
-    public class Moon
-    {
-        public int Id { get; set; }
-        
-        public string Name { get; set; }
-    }
-}
