--- src/DynamoApi/Startup.cs
@@ -1,61 +1,0 @@
-using Amazon.DynamoDBv2;
-using DynamoApi.HealthChecks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.OpenApi.Models;
-using Serilog;
-
-namespace DynamoApi
-{
-    public class Startup
-    {
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddControllers();
-            
-            services.AddHealthChecks()
-                .AddCheck<AliveHealthCheck>("alive_health_check")
-                .AddCheck<DynamoHealthCheck>("dynamo_health_check");
-            
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new OpenApiInfo { Title = "Example API", Version = "v1" });
-            });
-            
-            var dynamoDbConfig = Program.Configuration.GetSection("DynamoDb");
-            services.AddSingleton<IAmazonDynamoDB>(sp =>
-            {
-                var clientConfig = new AmazonDynamoDBConfig { ServiceURL = 
-                    dynamoDbConfig.GetValue<string>("LocalServiceUrl") };
-                return new AmazonDynamoDBClient(clientConfig);
-            });
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            
-            app.UseSerilogRequestLogging();
-
-            app.UseSwagger();
-            app.UseSwaggerUI(c =>
-            {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "Example API v1");
-            });
-
-            app.UseRouting();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-                endpoints.MapHealthChecks("/health");
-            });
-        }
-    }
-}
