--- src/DynamoApi/appsettings.json
@@ -1,19 +1,0 @@
-{
-  "Serilog": {
-    "MinimumLevel": {
-      "Default": "Information",
-      "Override": {
-        "Microsoft": "Warning",
-        "System": "Warning"
-      }
-    }
-  },
-  "AllowedHosts": "*",
-  "AWS": {
-    "Region": "us-east-1"
-  },
-  "DynamoDb": {
-    "LocalMode": true,
-    "LocalServiceUrl": "http://localhost:4569"
-  }
-}
