+++ src/Solar.Core/Entities/BaseEntity.cs
@@ -1,0 +1,7 @@
+namespace Solar.Core.Entities
+{
+    public abstract class BaseEntity
+    {
+        public int Id { get; protected set; }   
+    }
+}
