+++ src/Solar.Core/Entities/Planet.cs
@@ -1,0 +1,8 @@
+namespace Solar.Core.Entities
+{
+    public class Planet : BaseEntity
+    {
+        public string Universe { get; set; }
+        public string Name { get; set; }
+    }
+}
