+++ src/Solar.Core/Interfaces/IAsyncRepository.cs
@@ -1,0 +1,9 @@
+using Solar.Core.Entities;
+
+namespace Solar.Core.Interfaces
+{
+    // https://raw.githubusercontent.com/dotnet-architecture/eShopOnWeb/master/src/ApplicationCore/Interfaces/IAsyncRepository.cs
+    public interface IAsyncRepository<T> : IReadRepository<T>, IWriteRepository<T> where T : BaseEntity
+    {
+    }
+}
