+++ src/Solar.Web/Controllers/VersionController.cs
@@ -1,0 +1,21 @@
+using System.Reflection;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Solar.Web.Controllers
+{
+    [ApiController]
+    [Route("/version")]
+    public class VersionController : ControllerBase
+    {
+        [HttpGet]
+        public IActionResult GetVersion()
+        {
+            // ReSharper disable once AssignNullToNotNullAttribute
+            var version = Assembly.GetEntryAssembly()
+                .GetCustomAttribute<AssemblyInformationalVersionAttribute>()
+                ?.InformationalVersion;
+            
+            return Ok(new { Version = version });
+        }
+    }
+}
