+++ src/Solar.Web/HealthChecks/DynamoHealthCheck.cs
@@ -1,0 +1,16 @@
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+
+namespace Solar.Web.HealthChecks
+{
+    public class DynamoHealthCheck : IHealthCheck
+    {
+        public Task<HealthCheckResult> CheckHealthAsync(
+            HealthCheckContext context, 
+            CancellationToken cancellationToken = new CancellationToken())
+        {
+            return Task.FromResult(HealthCheckResult.Healthy("A healthy result."));
+        }
+    }
+}
