+++ src/Solar.Web/appsettings.Development.json
