+++ src/Solar.Web/appsettings.json
@@ -1,0 +1,16 @@
+{
+  "Serilog": {
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Warning",
+        "System": "Warning"
+      }
+    }
+  },
+  "AllowedHosts": "*",
+  "DynamoDb": {
+    "LocalMode": true,
+    "LocalServiceUrl": "http://localhost:4569"
+  }
+}
