--- tests/DynamoApi.IntegrationTests/TestCollection.cs
@@ -1,9 +1,0 @@
-using Xunit;
-
-namespace DynamoApi.IntegrationTests
-{
-    [CollectionDefinition("Test Collection", DisableParallelization = true)]
-    public class TestCollection : ICollectionFixture<TestFixture>
-    {
-    }
-}
