+++ tests/Solar.IntegrationTests/TestCollection.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace Solar.IntegrationTests
+{
+    [CollectionDefinition("Test Collection", DisableParallelization = true)]
+    public class TestCollection : ICollectionFixture<TestFixture>
+    {
+    }
+}
